Create mac_query.sh
Assign VLAN based on CALLING_STATION_ID attribute (MAC address)
when a corresponding entry is found in "mactab" file.

This is an external BASH script, called from freeradius. (must be configured for this in freeradius configs)
Works for cisco wireless lan controllers (WLCs) and Ruckus ZoneDirectors

This script is based on example code found on a freeradius mailing list that I've been unable to find/locate lately.